remove code repetitions
