+++ hedera-mirror-web3/src/test/java/com/hedera/mirror/web3/service/AbstractContractCallServiceTest.java
@@ -34,6 +34,7 @@ import com.hedera.mirror.web3.exception.MirrorEvmTransactionException;
 import com.hedera.mirror.web3.service.model.CallServiceParameters.CallType;
 import com.hedera.mirror.web3.service.model.ContractDebugParameters;
 import com.hedera.mirror.web3.service.model.ContractExecutionParameters;
+import com.hedera.mirror.web3.state.MirrorNodeState;
 import com.hedera.mirror.web3.utils.ContractFunctionProviderRecord;
 import com.hedera.mirror.web3.viewmodel.BlockType;
 import com.hedera.mirror.web3.web3j.TestWeb3jService;
@@ -43,9 +44,15 @@ import com.hedera.services.store.models.Id;
 import com.hedera.services.utils.EntityIdUtils;
 import com.hederahashgraph.api.proto.java.Key;
 import com.swirlds.state.State;
+import jakarta.annotation.PostConstruct;
 import jakarta.annotation.Resource;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
 import java.math.BigInteger;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.function.Consumer;
 import org.apache.commons.lang3.tuple.Pair;
@@ -591,6 +598,28 @@ public abstract class AbstractContractCallServiceTest extends Web3IntegrationTes
 
     protected String getAddressFromEvmAddress(final byte[] evmAddress) {
         return Address.wrap(Bytes.wrap(evmAddress)).toHexString();
+    }
+
+    protected void activateModularizedFlagAndInitializeState()
+            throws InvocationTargetException, IllegalAccessException {
+        mirrorNodeEvmProperties.setModularizedServices(true);
+
+        Method postConstructMethod = Arrays.stream(MirrorNodeState.class.getDeclaredMethods())
+                .filter(method -> method.isAnnotationPresent(PostConstruct.class))
+                .findFirst()
+                .orElseThrow(() -> new IllegalStateException("@PostConstruct method not found"));
+
+        postConstructMethod.setAccessible(true);
+        postConstructMethod.invoke(state);
+
+        final Map<String, String> propertiesMap = new ConcurrentHashMap<>();
+        propertiesMap.put("contracts.maxRefundPercentOfGasLimit", "100");
+        propertiesMap.put("contracts.maxGasPerSec", "15000000");
+        mirrorNodeEvmProperties.setProperties(propertiesMap);
+    }
+
+    protected void deactivateModularizedFlag() {
+        mirrorNodeEvmProperties.setModularizedServices(false);
     }
 
     public enum KeyType {
